requirements.txt
streamlit>=1.35.0
python-dotenv
Pillow
requests
fpdf2
beautifulsoup4
urllib3
googlesearch-python